% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iforest.R
\name{isolationForest.formula}
\alias{isolationForest.formula}
\title{Isolation forest formula}
\usage{
\method{isolationForest}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{Formula}

\item{data}{input data}
}
\value{
Isolation forest object
}
\description{
This implements formula S3 method.
}
