% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iforest.R
\name{isolationForestBuild}
\alias{isolationForestBuild}
\title{Builds trees for isolation forest}
\usage{
isolationForestBuild(X, num_trees, sample_size, ncols_per_tree, max_depth)
}
\arguments{
\item{X}{Dataframe of predictors}

\item{num_trees}{Number of trees in forest}

\item{sample_size}{Number of sampled rows to create one tree}

\item{ncols_per_tree}{Number of sampled columns to create one tree}

\item{max_depth}{Maximal depth of tree}
}
\value{
list with vector of isolation trees and parameters
}
\description{
This builds trees for isolation forest based on parameters.
}
