% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iforest.R
\name{path_length}
\alias{path_length}
\title{Calculates the path length of sample}
\usage{
path_length(x, T, curr_length = 0)
}
\arguments{
\item{x}{data sample}

\item{T}{isolation tree}

\item{curr_length}{current length of path}
}
\value{
length of path from root to external node
}
\description{
This calculates the path length of sample in tree
}
