% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iforest.R
\name{predict.isolationForest}
\alias{predict.isolationForest}
\title{Predicts method for isolation forest}
\usage{
\method{predict}{isolationForest}(object, newdata, ...)
}
\arguments{
\item{object}{isolation forest model}

\item{newdata}{data to predict}
}
\value{
array of predicted values from range (0,1)
}
\description{
This predicts values on new data using isolation forest.
}
