% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iforest.R
\name{summary.isolationForest}
\alias{summary.isolationForest}
\title{Prints summary information of isolation forest}
\usage{
\method{summary}{isolationForest}(X, ...)
}
\arguments{
\item{X}{Isolation forest object}
}
\description{
This prints parameters of created isolation forest.
}
